# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: MOSTLYVERS API
  version: 1.0.0
  description: Contract for the Android reader and Owner dashboard. FREE_LAUNCH grants non-financial entitlements; payment routes remain reserved and fail closed.
servers:
  - url: http://localhost:5001/v1
  - url: https://api.example.com/v1
...
paths:
...
  /me/books: { get: { tags: [Reading], operationId: ownedBooks, security: [ { bearerAuth: [] } ], responses: { '200': { $ref: '#/components/responses/Page' } } } }
  /books/{bookId}/reading-sessions: { parameters: [ { $ref: '#/components/parameters/BookId' } ], post: { tags: [Reading], operationId: createReadingSession, security: [ { bearerAuth: [] } ], responses: { '201': { $ref: '#/components/responses/Object' }, default: { $ref: '#/components/responses/Error' } } } }
  /reading-sessions/{sessionId}/resources/{resourceId}: { parameters: [ {name: sessionId, in: path, required: true, schema: {type: string, format: uuid}}, {name: resourceId, in: path, required: true, schema: {type: string}} ], get: { tags: [Reading], operationId: readResource, security: [ { bearerAuth: [] } ], responses: { '200': { description: Protected EPUB resource }, default: { $ref: '#/components/responses/Error' } } } }
  /me/books/{bookId}/progress:
    parameters: [ { $ref: '#/components/parameters/BookId' } ]
    get: { tags: [Reading], operationId: getProgress, security: [ { bearerAuth: [] } ], responses: { '200': { $ref: '#/components/responses/Object' }, '204': {description: No progress} } }
    put: { tags: [Reading], operationId: putProgress, security: [ { bearerAuth: [] } ], responses: { '200': { $ref: '#/components/responses/Object' }, default: { $ref: '#/components/responses/Error' } } }
  /me/books/{bookId}/bookmark:
    parameters: [ { $ref: '#/components/parameters/BookId' } ]
    put: { tags: [Reading], operationId: putBookmark, security: [ { bearerAuth: [] } ], responses: { '200': { $ref: '#/components/responses/Object' } } }
    delete: { tags: [Reading], operationId: deleteBookmark, security: [ { bearerAuth: [] } ], responses: { '204': {description: Removed} } }
...
  /me/books/{bookId}/offline-license/renew: { parameters: [ { $ref: '#/components/parameters/BookId' } ], post: { tags: [Reading], operationId: renewOfflineLicense, security: [ { bearerAuth: [] } ], responses: { '200': { $ref: '#/components/responses/Object' } } } }
...
  /me/downloads:
    get: { tags: [Reading], operationId: downloads, security: [ { bearerAuth: [] } ], responses: { '200': { $ref: '#/components/responses/Page' } } }
  /me/downloads/{bookId}: { parameters: [ { $ref: '#/components/parameters/BookId' } ], delete: { tags: [Reading], operationId: deleteDownload, security: [ { bearerAuth: [] } ], responses: { '204': {description: Local record removed} } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    internalToken: { type: http, scheme: bearer }
  parameters:
    BookId: { name: bookId, in: path, required: true, schema: {type: string, format: uuid} }
...
  responses:
    Object: { description: Success, content: { application/json: { schema: {type: object, additionalProperties: true} } } }
    Page: { description: Cursor page, content: { application/json: { schema: {$ref: '#/components/schemas/CursorPage'} } } }
...
    Error: { description: API error, content: { application/json: { schema: {$ref: '#/components/schemas/ErrorEnvelope'} } } }
...
  schemas:
...
    CursorPage: { type: object, required: [items], properties: {items: {type: array, items: {}}, nextCursor: {type: [string, 'null']}, total: {type: integer}} }
    ErrorEnvelope: { type: object, required: [error], properties: {error: {type: object, required: [code, message, requestId], properties: {code: {type: string}, message: {type: string}, requestId: {type: string}, details: {type: object, additionalProperties: true}}}} }